Simplify MoveToFront
This CL removes the templating from the MoveToFront code as all non-test
code uses uint32_t as the variable.
