+++ source/comp/CMakeLists.txt
@@ -13,7 +13,10 @@
 # limitations under the License.
 
 if(SPIRV_BUILD_COMPRESSION)
-  add_library(SPIRV-Tools-comp markv_codec.cpp)
+  add_library(SPIRV-Tools-comp
+      markv_codec.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/../util/move_to_front.h
+      ${CMAKE_CURRENT_SOURCE_DIR}/../util/move_to_front.cpp)
 
   spvtools_default_compile_options(SPIRV-Tools-comp)
   target_include_directories(SPIRV-Tools-comp
